+++ R/list_regions.R
@@ -3,7 +3,8 @@
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #' @param hl  Language used for text values. Optional. Default is \code{en-US}. For other allowed language codes, see \code{\link{list_langs}}.
 #' 
-#' @return data.frame with 3 columns: \code{gl} (two letter abbreviation), \code{name} (of the region), \code{etag}
+#' @return data.frame with 3 columns: 
+#' \code{gl} (two letter abbreviation), \code{name} (of the region), \code{etag}
 #' 
 #' @export
 #' 
@@ -23,17 +24,15 @@ list_regions <- function (hl = NULL, ...) {
 
 	res <- tuber_GET("i18nRegions", querylist, ...)
 
-	resdf <- NA
+	resdf <- read.table(text = "", col.names = c("gl", "name", "etag"))
+
+	# Cat total results
+	cat("Total Number of Content Regions:", length(res$items), "\n")
 
 	if (length(res$items) != 0) {
 		simple_res  <- lapply(res$items, function(x) c(unlist(x$snippet), etag=x$etag))
 		resdf       <- as.data.frame(do.call(rbind, simple_res))
-	} else {
-		resdf <- 0
 	}
-
-	# Cat total results
-	cat("Total Number of Content Regions:", length(res$items), "\n")
 
 	resdf
 }